+++ app/Http/Kernel.php
@@ -18,7 +18,9 @@ class Kernel extends HttpKernel
 		'Illuminate\View\Middleware\ShareErrorsFromSession',
 		'Illuminate\Foundation\Http\Middleware\VerifyCsrfToken',
 	    'MyBB\Settings\Middleware\SaveSettingsOnTerminate',
-		'MyBB\Core\Http\Middleware\BootstrapMyBB',
+		'MyBB\Core\Http\Middleware\SetupLanguage',
+		'MyBB\Core\Http\Middleware\ShareViewVariables',
+		'MyBB\Core\Http\Middleware\UpdateLastVisit',
 	];
 
 	/**
